+++ Src/Pscx.WinAdmin/Commands/Database/InvokeOleDbCommand.cs
@@ -13,7 +13,7 @@ using System.Management.Automation;
 
 namespace Pscx.Commands.Database
 {
-    [Cmdlet(PscxWinVerbs.Invoke, PscxWinNouns.OleDbCommand, SupportsShouldProcess=true), Description("Invoke commands on OleDb datasources")]
+    [Cmdlet(VerbsLifecycle.Invoke, PscxWinAdminNouns.OleDbCommand, SupportsShouldProcess=true), Description("Invoke commands on OleDb datasources")]
     public class InvokeOleDbCommand : OleDbCommandBase
     {
         protected override void ProcessRecord()