+++ Src/Pscx.WinAdmin/PscxWinAdminNouns.cs
@@ -1,0 +1,13 @@
+namespace Pscx
+{
+    internal static class PscxWinAdminNouns
+    {
+        public const string AdoCommand = "AdoCommand";
+        public const string AdoConnection = "AdoConnection";
+        public const string AdoDataProvider = "AdoDataProvider";
+        public const string ForegroundWindow = "ForegroundWindow";
+        public const string OleDbCommand = "OleDbCommand";
+        public const string OleDbData = "OleDbData";
+        public const string OleDbDataSet = "OleDbDataSet";
+    }
+}